14.6. Implementando serviço de cadastro de foto de produto

+++ src/main/java/com/algaworks/algafood/domain/model/FotoProduto.java
@@ -36,7 +36,7 @@ public class FotoProduto {
     private String descricao;
     @NotNull
     private String contentType;
-    private Integer tamanho;
+    private Long tamanho;
 
 
 }
